+++ src/main/java/web/dao/UserDao.java
@@ -10,9 +10,9 @@ public interface UserDao {
 
     void createUser(User user);
 
-    User showUser(int id);
+    User showUser(Long id);
 
-    void deleteUser(int id);
+    void deleteUser(Long id);
 
     void updateUser(User user);
